Update HOSTING.md
This is a tiny update. It was unclear to me what `ozone@example.com` should be replaced with. In my case with the domain `ozone.procycling.social`, replacing it with `ozone@procycling.social` worked, but I'm not sure what it should be replaced with if your domain is a root domain for example.